project15_nql_new/project15_nql_new/requirements.txt
alabaster==0.7.12
amqp==2.2.2
asn1crypto==0.24.0
autocommand==2.2.1
Babel==2.7.0
billiard==3.5.0.3
celery==4.2.1
certifi==2018.4.16
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
cycler==0.10.0
Django==2.2
django-cors-middleware==1.3.1
django-daterange-filter
django-hardcopy==0.1.4
django-oauth-toolkit==1.2.0
django-proxy==1.2.1
djangorestframework==3.9.4
docutils==0.15.2
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
et-xmlfile==1.0.1
fabric~=2.6.0
faiss~=1.5.3
faiss-cpu==1.6.1
filetype==1.0.7
idna==2.6
imagesize==1.1.0
importlib-metadata==0.8
importlib-resources==1.0.2
ipaddress==1.0.22
itsdangerous~=1.1.0
jdcal==1.3
jieba==0.42.1
Jinja2==2.10.3
joblib~=1.0.1
kiwisolver==1.0.1
kombu==4.2.1
lxml~=4.5.1
MarkupSafe==1.1.1
matplotlib~=3.2.2
mongo-connector==2.5.1
mongoengine~=0.23.0
mongomock==3.18.0
Naked==0.1.31
numpy~=1.19.0
oauthlib==2.0.7
openpyxl==2.5.0
packaging==19.2
patool==1.12
Pillow~=8.2.0
psycopg2-binary==2.7.6.1
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.6.1
Pygments==2.4.2
pylibmc==1.6.0
pymongo~=3.11.3
pyOpenSSL==19.0.0
pyparsing==2.3.1
python-dateutil==2.8.0
python-memcached==1.59
pandas==1.1.5
pytz~=2021.1
PyYAML==5.1
qrcode==6.1
requests~=2.25.1
scikit-learn==0.21.3
scipy~=1.5.0
sentinels==1.0.0
sentry-sdk==0.12.3
shellescape==3.4.1
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
treelib~=1.6.1
urllib3==1.22
vine==1.1.4
webservices==0.7
zipp==0.3.3
uwsgi
antlr4-python3-runtime==4.9.3